authors: add custom validation for date types
INSPIR-2770
