+++ src/app/shared/config/commons.ts
@@ -20,7 +20,7 @@
  * as an Intergovernmental Organization or submit itself to any jurisdiction.
  */
 
-import { RefAnchorAttributes, JsonStoreService, KeysStoreService, AutocompletionConfig } from 'ng2-json-editor';
+import { RefAnchorAttributes, JsonStoreService, KeysStoreService, AutocompletionConfig, CustomFormatValidation } from 'ng2-json-editor';
 
 import { environment } from '../../../environments/environment';
 import { ISO_LANGUAGE_MAP } from '../../shared/constants';
@@ -63,6 +63,33 @@ export const journalTitleAutocompletionConfig: AutocompletionConfig = {
   }
 };
 
+export const customValidationForDateTypes: CustomFormatValidation = {
+  date: {
+    formatChecker: (value) => {
+      let formats = [
+        /^\d{4}$/,
+        /^\d{4}-\d{2}$/,
+        /^\d{4}-\d{2}-\d{2}$/
+      ];
+      return formats
+        .some(format => {
+          if (value.match(format)) {
+            return Date.parse(value) !== NaN;
+          }
+          return false;
+        });
+    }
+  },
+  'date-time': {
+    formatChecker: (value) => {
+      let regex = /^\d\d\d\d-[0-1]\d-[0-3]\d[t\s][0-2]\d:[0-5]\d:[0-5]\d(?:\.\d+)?(?:z|[+-]\d\d:\d\d)?$/i;
+      if (value.match(regex)) {
+        return true;
+      }
+      return false;
+    }
+  }
+};
 
 export function fullTextSearch(value: any, expression: string): boolean {
   return JSON.stringify(value).search(new RegExp(expression, 'i')) > -1;
