+++ src/app/shared/config/index.ts
@@ -1,4 +1,5 @@
 import { hep, thesis, book, bookChapter, proceedings, conferencePaper } from './hep';
+import { authors } from './authors.config';
 
 export const editorConfigs = {
   hep,
@@ -6,7 +7,8 @@ export const editorConfigs = {
   book,
   'book chapter': bookChapter,
   proceedings,
-  'conference paper': conferencePaper
+  'conference paper': conferencePaper,
+  authors,
 };
 
 export * from './api.config';
